player controller and playerImpl commit

+++ src/main/java/terrato/springframwork/service/implementation/LeagueServiceImpl.java
@@ -48,7 +48,6 @@ public class LeagueServiceImpl implements LeagueService {
     }
 
 
-
     @Override
     @Transactional
     public League saveLeague(League league) {
@@ -61,10 +60,14 @@ public class LeagueServiceImpl implements LeagueService {
     @Override
     public void deleteLeagueById(Long idLeague) {
         Optional<League> leagueOptional = Optional.ofNullable(leagueRepository.findOne(idLeague));
-        League league1 = leagueOptional.get();
-        leagueRepository.delete(league1);
-    }
 
+        if (!leagueOptional.isPresent()) {
+            League league1 = leagueOptional.get();
+            leagueRepository.delete(league1);
+        } else {
+            throw new RuntimeException("I can't find league.");
+        }
+    }
 
 
 }
